move utils, schemas and views in right locations
