+++ chord_metadata_service/restapi/fhir_utils.py
@@ -81,7 +81,7 @@ def check_disease_onset(disease):
     return False
 
 
-##################### Class-based FHIR conversion functions #####################
+##################### Phenopackets to FHIR class conversion functions #####################
 
 
 def fhir_patient(obj):
@@ -384,3 +384,119 @@ def fhir_composition(obj):
             composition.section.append(section_content)
 
     return composition.as_json()
+
+
+##################### FHIR to Phenopackets class conversion functions #####################
+# There is no guide to map FHIR to Phenopackets
+
+def patient_to_individual(obj):
+    """ FHIR Patient to Individual. """
+
+    patient = p.Patient(obj)
+    individual = {
+        "id": patient.id
+    }
+    if patient.identifier:
+        individual["alternate_ids"] = [alternate_id.value for alternate_id in patient.identifier]
+    gender_to_sex = {
+        "male": "MALE",
+        "female": "FEMALE",
+        "other": "OTHER_SEX",
+        "unknown": "UNKNOWN_SEX"
+    }
+    if patient.gender:
+        individual["sex"] = gender_to_sex.get(patient.gender, "unknown")
+    if patient.birthDate:
+        individual["date_of_birth"] = patient.birthDate.isostring
+    if patient.active:
+        individual["active"] = patient.active
+    if patient.deceasedBoolean:
+        individual["deceased"] = patient.deceasedBoolean
+    return individual
+
+
+def observation_to_phenotypic_feature(obj):
+    """ FHIR Observation to Phenopackets PhenotypicFeature. """
+
+    observation = obs.Observation(obj)
+    codeable_concept = observation.code  # CodeableConcept
+    phenotypic_feature = {
+        # id is an integer AutoField, store legacy id in description
+        # TODO change
+        "description": observation.id,
+        "pftype": {
+            "id": codeable_concept.coding[0].code,
+            "label": codeable_concept.coding[0].display
+            # TODO collect system info in metadata
+        }
+    }
+    if observation.specimen:  # FK to Biosample
+        phenotypic_feature["biosample"] = observation.specimen.reference
+    return phenotypic_feature
+
+
+def condition_to_disease(obj):
+    """ FHIR Condition to Phenopackets Disease. """
+
+    condition = cond.Condition(obj)
+    codeable_concept = condition.code  # CodeableConcept
+    disease = {
+        # id is an integer AutoField, legacy id can be a string
+        # "id": condition.id,
+        "term": {
+            "id": codeable_concept.coding[0].code,
+            "label": codeable_concept.coding[0].display
+            # TODO collect system info in metadata
+        }
+    }
+    # condition.stage.type is only in FHIR 4.0.0 version
+    return disease
+
+
+def diagnostic_report_to_interpretation(obj):
+    """ FHIR DiagnosticReport to Phenopackets Interpretation. """
+    # it hardly maps at all
+    return
+
+
+def procedure_to_procedure(obj):
+    """ FHIR Procedure to Phenopackets Procedure.
+    The main semantic difference:
+    - phenopackets procedure is a procedure performed to extract a biosample;
+    - fhir procedure is a procedure performed on or for a patient
+    (e.g. documentation of patient's medication)
+    """
+
+    return
+
+
+def specimen_to_biosample(obj):
+    """ FHIR Specimen to Phenopackets Biosample. """
+
+    specimen = s.Specimen(obj)
+    biosample = {
+        "id": specimen.id
+    }
+    if specimen.subject:
+        biosample["individual"] = specimen.subject.reference
+    if specimen.type:
+        codeable_concept = specimen.type  # CodeableConcept
+        biosample["sampled_tissue"] = {
+            "id": codeable_concept.coding[0].code,
+            "label": codeable_concept.coding[0].display
+            # TODO collect system info in metadata
+        }
+    if specimen.collection:
+        method_codeable_concept = specimen.collection.method
+        bodysite_codeable_concept = specimen.collection.bodySite
+        biosample["procedure"] = {
+            "code": {
+                "id": method_codeable_concept.coding[0].code,
+                "label": method_codeable_concept.coding[0].display
+            },
+            "body_site": {
+                "id": bodysite_codeable_concept.coding[0].code,
+                "label": bodysite_codeable_concept.coding[0].display
+            }
+        }
+    return biosample
