+++ chord_metadata_service/restapi/views_ingest_fhir.py
@@ -2,7 +2,13 @@ import json
 import uuid
 import jsonschema
 
-from fhirclient.models import observation as obs, patient as p, condition as cond, specimen as s
+from .schemas import FHIR_INGEST_SCHEMA, FHIR_BUNDLE_SCHEMA
+from .fhir_utils import (
+    patient_to_individual,
+    observation_to_phenotypic_feature,
+    condition_to_disease,
+    specimen_to_biosample
+)
 
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import AllowAny
@@ -12,58 +18,16 @@ from chord_lib.responses.errors import *
 from chord_metadata_service.chord.models import *
 from chord_metadata_service.phenopackets.models import *
 
-FHIR_INGEST_SCHEMA = {
-    "$id": "chord_metadata_service_fhir_ingest_schema",
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "description": "FHIR Ingest schema",
-    "type": "object",
-    "properties": {
-        "dataset_id": {"type": "string"},
-        "patients": {"type": "string", "description": "Path to a patients file location."},
-        "observations": {"type": "string", "description": "Path to an observations file location."},
-        "conditions": {"type": "string", "description": "Path to a conditions file location."},
-        "specimens": {"type": "string", "description": "Path to a specimens file location."},
-        "metadata": {
-            "type": "object",
-            "properties": {
-                "created_by": {"type": "string"}
-            },
-            "required": ["created_by"]
-        }
-    },
-    "required": [
-        "dataset_id",
-        "patients",
-        "metadata"
-    ],
-}
 
-
-FHIR_BUNDLE_SCHEMA = {
-    "$id": "chord_metadata_service_fhir_bundle_schema",
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "description": "FHIR Bundle schema",
-    "type": "object",
-    "properties": {
-        "resourceType": {
-            "type": "string",
-            "const": "Bundle",
-            "description": "Collection of resources."
-        },
-        "entry": {
-            "type": "array",
-            "items": {
-                "type": "object",
-                "properties": {
-                    "resource": {"type": "object"}
-                },
-                "additionalProperties": True,
-                "required": ["resource"]
-            }
-        }
-    },
-    "additionalProperties": True,
-    "required": ["resourceType", "entry"]
+INGEST_BODY_EXAMPLE = {
+    "dataset_id": "62b5fc67-d925-4409-bb59-e1e9a1ef10ae",
+    "patients": "examples/patients.json",
+    "observations": "examples/observations.json",
+    "conditions": "examples/conditions.json",
+    "specimens": "examples/specimens.json",
+    "metadata": {
+        "created_by": "Ksenia Zaytseva"
+    }
 }
 
 
@@ -82,7 +46,7 @@ def _check_schema(schema, obj, additional_info=None):
         error_messages = []
         for i, error in enumerate(errors, 1):
             error_messages.append(f"{i} validation error {'.'.join(str(v) for v in error.path)}: {error.message}")
-        raise ValidationError(f"{additional_info+' ' if additional_info else None}errors: {error_messages}")
+        raise ValidationError(f"{additional_info + ' ' if additional_info else None}errors: {error_messages}")
 
 
 @api_view(["POST"])
@@ -158,7 +122,7 @@ def ingest_fhir(request):
                 _check_schema(FHIR_BUNDLE_SCHEMA, conditions_data, 'conditions data')
                 for item in conditions_data["entry"]:
                     disease_data = condition_to_disease(item["resource"])
-                    disease= Disease.objects.create(**disease_data)
+                    disease = Disease.objects.create(**disease_data)
                     # Condition must have a subject
                     if not item["resource"]["subject"]:
                         return Response(bad_request_error(f"Subject is required."), status=404)
@@ -195,119 +159,3 @@ def ingest_fhir(request):
                 return Response(bad_request_error(f"Invalid JSON provided (message: {e})"), status=400)
 
     return Response(status=204)
-
-
-################################# FHIR to Phenopackets #################################
-# There is no guide to map FHIR to Phenopackets
-
-def patient_to_individual(obj):
-    """ FHIR Patient to Individual. """
-
-    patient = p.Patient(obj)
-    individual = {
-        "id": patient.id
-    }
-    if patient.identifier:
-        individual["alternate_ids"] = [alternate_id.value for alternate_id in patient.identifier]
-    gender_to_sex = {
-        "male": "MALE",
-        "female": "FEMALE",
-        "other": "OTHER_SEX",
-        "unknown": "UNKNOWN_SEX"
-    }
-    if patient.gender:
-        individual["sex"] = gender_to_sex.get(patient.gender, "unknown")
-    if patient.birthDate:
-        individual["date_of_birth"] = patient.birthDate.isostring
-    if patient.active:
-        individual["active"] = patient.active
-    if patient.deceasedBoolean:
-        individual["deceased"] = patient.deceasedBoolean
-    return individual
-
-
-def observation_to_phenotypic_feature(obj):
-    """ FHIR Observation to Phenopackets PhenotypicFeature. """
-
-    observation = obs.Observation(obj)
-    codeable_concept = observation.code  # CodeableConcept
-    phenotypic_feature = {
-        # id is an integer AutoField, store legacy id in description
-        # TODO change
-        "description": observation.id,
-        "pftype": {
-            "id": codeable_concept.coding[0].code,
-            "label": codeable_concept.coding[0].display
-            # TODO collect system info in metadata
-        }
-    }
-    if observation.specimen:  # FK to Biosample
-        phenotypic_feature["biosample"] = observation.specimen.reference
-    return phenotypic_feature
-
-
-def condition_to_disease(obj):
-    """ FHIR Condition to Phenopackets Disease. """
-
-    condition = cond.Condition(obj)
-    codeable_concept = condition.code  # CodeableConcept
-    disease = {
-        # id is an integer AutoField, legacy id can be a string
-        # "id": condition.id,
-        "term": {
-            "id": codeable_concept.coding[0].code,
-            "label": codeable_concept.coding[0].display
-            # TODO collect system info in metadata
-        }
-    }
-    # condition.stage.type is only in FHIR 4.0.0 version
-    return disease
-
-
-def diagnostic_report_to_interpretation(obj):
-    """ FHIR DiagnosticReport to Phenopackets Interpretation. """
-    # it hardly maps at all
-    return
-
-
-def procedure_to_procedure(obj):
-    """ FHIR Procedure to Phenopackets Procedure.
-    The main semantic difference:
-    - phenopackets procedure is a procedure performed to extract a biosample;
-    - fhir procedure is a procedure performed on or for a patient
-    (e.g. documentation of patient's medication)
-    """
-
-    return
-
-
-def specimen_to_biosample(obj):
-    """ FHIR Specimen to Phenopackets Biosample. """
-
-    specimen = s.Specimen(obj)
-    biosample = {
-        "id": specimen.id
-    }
-    if specimen.subject:
-        biosample["individual"] = specimen.subject.reference
-    if specimen.type:
-        codeable_concept = specimen.type  # CodeableConcept
-        biosample["sampled_tissue"] = {
-            "id": codeable_concept.coding[0].code,
-            "label": codeable_concept.coding[0].display
-            # TODO collect system info in metadata
-        }
-    if specimen.collection:
-        method_codeable_concept = specimen.collection.method
-        bodysite_codeable_concept = specimen.collection.bodySite
-        biosample["procedure"] = {
-            "code": {
-                "id": method_codeable_concept.coding[0].code,
-                "label": method_codeable_concept.coding[0].display
-            },
-            "body_site": {
-                "id": bodysite_codeable_concept.coding[0].code,
-                "label": bodysite_codeable_concept.coding[0].display
-            }
-        }
-    return biosample
